Add a service layer for recipes
