+++ app/recipe/serializers.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 
 from core.models import Recipe, Ingredient
+from recipe.service import RecipeService
 
 
 class IngredientSerializer(serializers.ModelSerializer):
@@ -16,37 +17,19 @@ class RecipeSerializer(serializers.ModelSerializer):
     """Serializer for recipe objects"""
 
     ingredients = IngredientSerializer(many=True, required=False)
+    recipeService = RecipeService()
 
     class Meta:
         model = Recipe
         fields = ["id", "title", "ingredients"]
         read_only_fields = ["id"]
 
-    def _get_or_create_ingredients(self, ingredients, recipe):
-        for ingredient in ingredients:
-            ingredient_obj, _ = Ingredient.objects.get_or_create(**ingredient)
-            recipe.ingredients.add(ingredient_obj)
-
     def create(self, validated_data):
         """Create a recipe"""
-        ingredients = validated_data.pop("ingredients", [])
-        recipe = Recipe.objects.create(**validated_data)
-        self._get_or_create_ingredients(ingredients, recipe)
-        return recipe
+        return self.recipeService.create(validated_data)
 
     def update(self, instance, validated_data):
-        ingredients = validated_data.pop("ingredients", None)
-
-        if ingredients is not None:
-            instance.ingredients.clear()
-            self._get_or_create_ingredients(ingredients, instance)
-
-        for attr, value in validated_data.items():
-            setattr(instance, attr, value)
-
-        instance.save()
-
-        return instance
+        return self.recipeService.update(instance, validated_data)
 
 
 class RecipeDetailSerializer(RecipeSerializer):
